refactor: improve code quality and consistency across multiple modules
Clean up and standardize code patterns throughout the codebase for better
maintainability and consistency.

Changes:

BanCommands:
- Store FCManager reference instead of BanManager directly
- Fix ban duration calculation for default vs custom time
- Add getDefaultBanDuration() helper in Settings
- Simplify checkban command (remove unused BAN_REMAINING display)
- Fix time formatting to use seconds correctly

BuildCommand:
- Fix whitespace and formatting issues

Placeholders:
- Use Settings.getEnabledMatchTypes() instead of hardcoded Arrays.asList(1,2,3,4,5)
- Rename variable 'q' to 'queue' for better readability
- Remove unused Arrays import

PlayerSettings:
- Initialize default settings from Settings config instead of hardcoded values
- Use Settings for kick_sound, goal_sound, particles, build_mode defaults
- Use Settings for default goal_celebration message
- Rename loop variable 'p' to 'particle' for clarity

This patch improves consistency with the centralized Settings system and
makes the code more maintainable by reducing hardcoded values and improving
variable naming conventions.

Version: 1.14.34

+++ src/main/java/io/github/divinerealms/footcube/tasks/TouchCleanupTask.java
@@ -76,6 +76,7 @@ public class TouchCleanupTask extends BaseTask {
       for (CubeTouchType type : touchesToRemove) {
         playerTouches.remove(type);
       }
+
       if (playerTouches.isEmpty()) {
         playersToRemove.add(playerId);
       }
@@ -91,12 +92,14 @@ public class TouchCleanupTask extends BaseTask {
     if (raised.isEmpty()) {
       return;
     }
+
     cubesToRemove.clear();
     for (Map.Entry<UUID, Long> entry : raised.entrySet()) {
-      if ((now - entry.getValue()) > Settings.KICK_COOLDOWN_RISE.asInt()) {
+      if ((now - entry.getValue()) > Settings.KICK_COOLDOWN_RISE.asLong()) {
         cubesToRemove.add(entry.getKey());
       }
     }
+
     for (UUID cubeId : cubesToRemove) {
       raised.remove(cubeId);
     }
